requirements.txt
google-maps-places>=0.1.0
python-dotenv>=1.0.0
tqdm>=4.66.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
extruct>=0.11.0
crawl4ai>=0.3.0
instructor>=0.4.0
openai>=1.0.0
pydantic>=2.0.0
pdf2image>=1.16.0
Pillow>=10.0.0
